Modified code without using split function
